requirements.txt
python-telegram-bot==22.0
requests==2.31.0
aiogram
storage
apscheduler
pytz